+++ src/main/java/br/com/johnatanbrayan/domain/PagamentoComCartao.java
@@ -5,13 +5,14 @@ import javax.persistence.Entity;
 import br.com.johnatanbrayan.domain.enums.EstadoPagamento;
 
 @Entity
-public class PagamentoComCartao extends Pagamento{
+public class PagamentoComCartao extends Pagamento {
 	private static final long serialVersionUID = 1L;
 	
 	private Integer numeroDeParcelas;
 	
 	public PagamentoComCartao() {}
-	public PagamentoComCartao(Integer id, EstadoPagamento estadoPagamento, Pedido pedido, Integer numeroDeParcelas) {
+	
+	public PagamentoComCartao(Long id, EstadoPagamento estadoPagamento, Pedido pedido, Integer numeroDeParcelas) {
 		super(id,estadoPagamento, pedido);
 		this.numeroDeParcelas = numeroDeParcelas;
 	}
