+++ src/main/java/br/com/johnatanbrayan/repository/CategoriaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.johnatanbrayan.domain.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria,Integer>{}
+public interface CategoriaRepository extends JpaRepository<Categoria,Long>{}
