+++ src/main/java/br/com/johnatanbrayan/repository/ClienteRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.johnatanbrayan.domain.Cliente;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer>{}
+public interface ClienteRepository extends JpaRepository<Cliente, Long>{}
