+++ src/main/java/br/com/johnatanbrayan/repository/PagamentoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.johnatanbrayan.domain.Pagamento;
 
-public interface PagamentoRepository extends JpaRepository<Pagamento,Integer>{}
+public interface PagamentoRepository extends JpaRepository<Pagamento,Long>{}
