+++ src/main/java/br/com/johnatanbrayan/repository/PedidoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.johnatanbrayan.domain.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido,Integer>{}
+public interface PedidoRepository extends JpaRepository<Pedido,Long>{}
