+++ src/main/java/br/com/johnatanbrayan/repository/ProdutoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.johnatanbrayan.domain.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto,Integer>{}
+public interface ProdutoRepository extends JpaRepository<Produto,Long>{}
