+++ src/main/java/br/com/johnatanbrayan/resources/CategoriaResource.java
@@ -21,7 +21,7 @@ public class CategoriaResource {
 	private CategoriaService categoriaService;
 	
 	@RequestMapping(value="/{id}", method=RequestMethod.GET)
-	public ResponseEntity<?> get(@PathVariable Integer id){
+	public ResponseEntity<?> get(@PathVariable Long id){
 		Categoria obj = categoriaService.find(id);
 		return ResponseEntity.ok().body(obj);
 	}
