+++ src/main/java/br/com/johnatanbrayan/service/CategoriaService.java
@@ -15,7 +15,7 @@ public class CategoriaService {
 	@Autowired
 	private CategoriaRepository categoriaRepository;
 	
-	public Categoria find(Integer id) {
+	public Categoria find(Long id) {
 		Optional<Categoria> obj = categoriaRepository.findById(id);
 		return obj.orElseThrow(() -> new ObjectNotFoundException(
 				"Objeto não encontrado! Id: "+id+" ,Tipo: "+Categoria.class.getName()));
