+++ src/main/java/br/com/johnatanbrayan/service/ClienteService.java
@@ -15,7 +15,7 @@ public class ClienteService {
 	@Autowired
 	private ClienteRepository clienteRepository;
 	
-	public Cliente find(Integer id) {
+	public Cliente find(Long id) {
 		Optional<Cliente> obj = clienteRepository.findById(id);
 		return obj.orElseThrow(() -> new ObjectNotFoundException("Objeto não encontrado: "+id+" ,Tipo: "+
 		Cliente.class.getName()));
